+++ R/predict.R
@@ -17,67 +17,69 @@
 # with this program; if not, write to the Free Software Foundation, Inc.,
 # 51 Franklin Street, Fifth Floor, Boston, MA 02110-1301 USA.
 
-#' Assigning New Points to a Clustering
-#'
-#' Predict the membership of a new point given a clustering.
-#'
-#' **For DBSCAN:** If a new point is in the eps neighborhood of a point in the clustering, the
-#' the label of that point is predicted, otherwise the prediction is `0` (noise).
-#'
-#' **For OPTICS:** Extracts a DBSCAN clustering and use predict for DBSCAN.
-#'
-#' **For HDBSCAN:** Find the k-nearest neighbors with the smallest mutual
-#' reachability distance to predict the label
-#'
-#' @name predict
-#' @aliases predict
-#' @family clustering functions
-#'
+#' @rdname dbscan
 #' @param object clustering object.
 #' @param data the data set used to create the clustering object.
 #' @param newdata new data points for which the cluster membership should be
 #' predicted.
-#' @param ... further arguments.
-NULL
-
-#' @rdname predict
 predict.dbscan_fast <- function (object, newdata, data, ...)
   .predict_frNN(newdata, data, object$cluster, object$eps, ...)
 
-#' @rdname predict
+#' @rdname optics
+#' @param object clustering object.
+#' @param data the data set used to create the clustering object.
+#' @param newdata new data points for which the cluster membership should be
+#' predicted.
 predict.optics <- function (object, newdata, data, ...) {
-  if (is.null(object$cluster))
-    stop("no extracted clustering available in object! run extractDBSCAN or extractXi first.")
+  if (is.null(object$cluster) ||
+      is.null(object$eps_cl) || is.na(object$eps_cl))
+    stop("no extracted clustering available in object! run extractDBSCAN() first.")
   .predict_frNN(newdata, data, object$cluster, object$eps_cl, ...)
 }
 
-
-#' @rdname predict
+#' @rdname hdbscan
+#' @param object clustering object.
+#' @param data the data set used to create the clustering object.
+#' @param newdata new data points for which the cluster membership should be
+#' predicted.
 predict.hdbscan <- function(object, newdata, data, ...) {
-  k <- object$minPts - 1
+  clusters <- object$cluster
 
-  knn <- kNN(data, newdata, k = k , sort = TRUE)
-  nn <- knn$id
-  nn_dist <- knn$dist
-  core_dist <- nn_dist[, k]
+  if (is.null(newdata))
+    return(clusters)
 
-  nn_mrd <- mrd_m(nn_dist, core_dist)
+  # don't use noise
+  coredist <- object$coredist[clusters != 0]
+  data <- data[clusters != 0,]
+  clusters <- clusters[clusters != 0]
 
-  nn_id <- apply(nn_mrd, MARGIN = 1, which.min)
-  object$cluster[nn[cbind(seq(nrow(nn)), nn_id)]]
+  # find minPts - 1 nearest neighbor
+  nns <- kNN(data, query = newdata, k = 1)
+
+  # choose cluster if dist <= coredist of that point
+  drop(ifelse(nns$dist > coredist[nns$id], 0L, clusters[nns$id]))
 }
 
+## find the cluster id of the closest NN in the eps neighborhood or return 0 otherwise.
 .predict_frNN <- function(newdata, data, clusters, eps, ...) {
   if (is.null(newdata))
     return(clusters)
 
-  # calculate the frNN between newdata and data (only keep entries for newdata)
-  nn <- frNN(rbind(data, newdata), eps = eps,
-    sort = TRUE, ...)$id[-(1:nrow(data))]
+  # don't use noise
+  data <- data[clusters != 0,]
+  clusters <- clusters[clusters != 0]
 
-  # get rid of indices of newdata and pick first neighbor
-  cl <- sapply(nn, function(x)
-    clusters[x[x <= nrow(data) & x > 0][1]])
-  cl[is.na(cl)] <- 0L
-  cl
+  # calculate the frNN between newdata and data (only keep entries for newdata)
+  nn <- frNN(data,
+    query = newdata,
+    eps = eps,
+    sort = TRUE,
+    ...)
+
+  sapply(nn$id, function(nns) {
+    if (length(nns) == 0)
+      0L
+    else
+      clusters[nns[1]]
+  })
 }